backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
email-validator>=2.0.0
Pillow>=10.0.0
aiofiles>=23.0.0
pytest>=7.3.1
httpx>=0.24.1
redis>=4.6.0
boto3>=1.28.0
markdown>=3.4.0
slugify>=0.0.1
